+++ src/renderer/src/main.tsx
@@ -5,9 +5,14 @@ import { createRoot } from 'react-dom/client'
 import { RouterProvider } from '@tanstack/react-router'
 
 import { router } from './router'
+import { UpdaterListener } from './components/updater-listener'
+import { ToastProvider } from './components/ui/toast'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <RouterProvider router={router} />
+    <ToastProvider>
+      <RouterProvider router={router} />
+      <UpdaterListener />
+    </ToastProvider>
   </StrictMode>
 )